Address PR comments * change ERROR to WARNING * move test * update docs
